family-fuel-dashboard/requirements.txt
duckdb>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.31.0
openfoodfacts>=5.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=8.0.0
